backend/requirements.txt
flask
flask-cors
llama_index==0.11.16
nltk==3.9.1
openpyxl==3.1.2
pandas==1.2.5
python-dotenv==1.0.1
Requests==2.32.3
